Update useInterval hook
